ProductManagement/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
botocore==1.20.78
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
jmespath==0.10.0
kombu==5.0.2
numpy==1.19.5
pandas==1.1.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
